split cards in components
